Improves responsiveness on starting; uses Holo theme
The way theming was handled has been revised so that the app doesn't
switch themes on start. Instead, there are now two activities - one for
the dialog box Share mode and one for the app mode.

Bumps SDK target version so that the appropriate theming and
compatibility features are enabled.

Adds minor formatting, documentation, and performance improvements.