Reworks the VisualHash; Breaking Change: new hash images
The previous VisualHash was not written properly. As such, it was using
only a fraction of the bits available to it.

This reworks the hash by adding a few more shapes and properly breaking
out the color bits. Unfortunately, this means that all the hashes will
look different than previously.

There are still 3 extra bits of information per each shape that aren't
being used, so that should probably be fixed somehow.