Moves GeneratedPasswordView style to xml; adds gen icon
This also attempts to simplify the theming by using an inset drawable in
the GeneratedPasswordView instead of attempting to simulate the whole
widget's look. Trying to do the latter is a fight against many different
themes.

This also adds a password generation icon in an attempt to class things
up a bit.