Eliminates Go button; adds copy button
As the passwords are now generated as the user types, the Go button is
mostly useless. Now there's a copy button instead (which also adds the
current domain to the remembered domains, if that feature's enabled).

Pressing the Enter key still triggers the Go action.

The new copy feature is in the Action Bar or the Options Menu, depending
on Android version.