Allows manual pasting of full URL into domain field.
For those who don't use the share/send intent and instead paste
a full URL from their browser into the domain field, this patch
extracts the domain from the URL and uses it. This behavior largely
matches that of other versions of SuperGenPass.

Untested (still learning how to build for Android).