Disables URL parsing when domain checking is disabled
One of the settings disables domain checking, so any keyword can be
used. Having it not parse the input as a URL fits into this scheme.
While it would be somewhat silly for a URL to be used as a keyword,
none-the-less it makes sense that disabling "domain checking" would also
disable "URL parsing".