Adds ability to scan salts from a QR code
Instead of having to type in a complex salt (which may limit the amount
of entropy used in it), this lets you create a QR code which you can
then later scan.

One idea for generating such a QR code is the following:

dd if=/dev/random bs=1 count=512 | base64 -w 0 | qrencode -o /tmp/key.png

The code can then be printed and kept in a physically secure location.
If you ever need to re-key SuperGenPass, just scan the same code again.