Adds first pass at PIN generation; adds junit project
This feature uses a slightly modified HOTP to generate a PIN based on
the domain and password entered. The core functionality should be
finished, but the algorithm is not yet stable. UI still needs to be
worked out.

Due to the nature of PINs, it goes to special lengths to avoid "bad"
PINs. These pins are things like "1234" and "1111" which would be very
unfortunate to be stuck with if it happened to be generated for a given
password/domain combination.

This also includes the first pass at a junit testing project.