Fixes bug with PIN generation; BREAKING CHANGE
The domain used during PIN generation was not being stripped of
sub-domains. As a result, the PIN generation is out of sync with other
SGP implementations and the intended design. This resolves this issue.

PINs generated prior to this fix that contain a sub-domain (e.g.
foo.example.org) can still be recreated by disabling domain checking in
the settings. From this change forward, "foo.example.org" will have the
same PIN as "bar.example.org" and "example.org". This matches SGP.